Update upload_manager.py
passing `max_workers` as parameter, if `None` `concurrent.futures.ThreadPoolExecutor` will use the system cpu number * 5. From the [doc](https://docs.python.org/3/library/concurrent.futures.html)

Changed in version 3.5: If max_workers is None or not given, it will default to the number of processors on the machine, multiplied by 5, assuming that ThreadPoolExecutor is often used to overlap I/O instead of CPU work and the number of workers should be higher than the number of workers for ProcessPoolExecutor.