Deconstruct Paths for COCO (#218)
Adds support for deconstructing paths when importing csvtags and coco filenames. 

Many file format that we support have now conscept of folders or paths, 
therefore we need to extract a path from the filename itself. 
Example, filename: "a/b/c.png" will now be seen as path: "a/b/" and filename: "c.png"